WebGUI::Image::Graph::Pie using GD::Graph::pie instead
currently working, passing tests, and generating reasonable output,
but I need to pull all of the logic unrelated to generating pie
charts and related to using GD instead of Imager out into its
own module with its own place in the inheritance hierarchy.
this commit is an intermediate step there.