rip off Image::Magick's arm and beat it to death with the bloody stump.  closes #11849.
See http://www.webgui.org/community/webgui-8/issues/11849

This migrates wG8 to Imager and GD::Graph.

Image::Magick formerly did thumbnailing, scaling/rotation for gallery
images, and plotting for the Poll asset's various optional graphic graphs
(HTML barchart output was and is the default).  This finishes the
migration of all of those things away (pending bugs, of course).
There are no Image::Magick references left in the code.

GD::Graph is only used by lib/WebGUI/Image/Graph/Pie.pm which is
currently only used by the the Poll asset.  Poll doesn't list graph
renderers that don't load, so GD::Graph can be considered optional.

One request we've had for years is to put WebGUI on CPAN, but we always
reached the conclusion that we had to get rid of Image::Magick first.  I think
the discussion went along the lines of, yes, we could put it on CPAN, and we
want to, but first we need to get rid of Image::Magick.  Doing the source
install process was too often held up on this, and putting stuff on CPAN
would encourage people to attempt that rather than doing the WRE.  Image::Magick
has been a thorn in the side on the scale of mod_perl, which 8 also eliminated.

Here's more information on the Poll asset from the WebGUI Content Manager's
Guide:  http://www.webgui.or...guide-wiki/poll

WebGUI::Image::Graph::Pie is now implemented on top of GD::Graph::pie.
The other ones were literally translated to plot using Imager instead of I::M.