tidy README.md source instructions, and update Docker instructions for volumes
So, Docker has this volumes thing that gives containers persistant storage
without having to "commit" them (which is good, because commiting adds another
layer on top of the stack of union filesystems, which is bad).  But the
user has to take care to start/stop the same container by name, so tell
them that.  Updating this after adding volumes to the Docker image:
VOLUME /var/lib/mysql
VOLUME /data/domains/www.example.com/public/uploads