mysql hasn't been enforcing "not null" constraints.
as explained here:  http://stackoverflow.com/questions/7264702/i-set-a-mysql-column-to-not-null-but-still-i-can-insert-an-empty-value

mysql has traditionally been in non-strict mode.  a whole pile of
"not null" clauses are completely ignored by the code.  for example,
VersionTag.pm in create() and new() will insert a 'new' record
without a name or several other "not null" values, which would
silently fail as $session->db->setRow() completely blackholes
errors.

so, remove all of the "not null" clauses that don't also specify
a default.