installer:  'dialog' package for Debian
either Debian doesn't come with it any more or that's just the way this
https://hub.docker.com/_/perl/ image came.  I'm going to cry so hard if
it turns out that all of these Debian problems are specific to this Docker
image.

No real reason for the 'dialog' package.  I thought maybe I could inject
a password into the dialog during Debian's MySQL config, but nope
(so same as Readline method).  Keeping it because it kind of looks nice.
But on the other hand, it doesn't render completely correctly.

This only comes up when we shell out to run apt, which we only do for
MySQL on Debian, and that commit should follow.