fetch and unpack webgui earlier and prompting by verbosity
previous commits broke stuff.
fetch wG and unpack earlier so that the directory structure is there
for when we try to install cpanm (cpanm got moved up when things
got changed to try to read the config file before trying to set
up the database).
also, run()'s noprompt arg is defaulted based on verbosity level.